feat(order-service): CRUD, RestClient and Docker

+++ product-service/src/main/java/com/microservices/ordersystem/product_service/model/Product.java
@@ -4,6 +4,8 @@ import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Entity
 @Getter @Setter
 public class Product {
@@ -13,14 +15,14 @@ public class Product {
     private String name;
     @Enumerated(EnumType.STRING)
     private Category category;
-    private Double price;
+    private BigDecimal price;
 
     @Enumerated(EnumType.STRING)
     private Status status;
 
     public Product() {}
 
-    public Product(String name, Category category, Double price, Status status) {
+    public Product(String name, Category category, BigDecimal price, Status status) {
         this.name = name;
         this.category = category;
         this.price = price;
